+++ sales_archive/wizards/__init__.py
@@ -1,0 +1,1 @@
+from . import sale_custom_report